Split summary page behaviour to share with pdf generation
The existing `confirm-tenants` behaviour was handling multiple separate concerns - particularly the rendering of tenant data out to the table, the handling of outbound edit links, and the reconcilation of tenant data when returning from an edit journey.

The first of these needs to be re-used for the data pdf sent on submission, so split that into a separate behaviour so it can be reused across both pages.
